# api/swagger.yml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: API de Domains
  description: Documentación de la API de Domains de la materia 75.43 Introducción a Sistemas Distribuidos

servers:
  - url: http://localhost:8080/api
    description: Servidor de desarrollo

# Paths servidos por la aplicación
paths:
  /domains/{domain}:
      get:
        operationId: "api.domains.obtener_uno"
        tags:
          - "Domains"
        summary: "IP del domain provisto"
        description: "Obtiene un IP particular de un hostname provisto"
        parameters:
          - in: path
            name: domain
            required: true
            schema:
              type: string
            description: El hostname del domain a obtener
        responses:
          200:
            description: "Ok"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Domain'
          404:
            description: "Not Found"
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /custom-domains/{domain}:
    delete:
      operationId: "api.domains.borrar"
      tags:
        - "Domains"
      summary: "Borrar un custom domains de los domains"
      description: "Borra un custom domain"
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
          description: El hostname del custom-domain a borrar
      responses:
        200:
          description: "Ok"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/OnlyDomain'
        404:
          description: "Not Found"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
    put:
      operationId: "api.domains.modificar"
      tags:
        - "Domains"
      summary: "Modificar un custom domain de los domains"
      description: "Modifica un custom domain si este existe"
      parameters:
        - in: path
          name: domain
          required: true
          schema:
            type: string
          description: El hostname del custom-domain a modificar
      requestBody:
        required: true
        x-body-name: 'domain'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnlyIp'
      responses:
        200:
          description: "Ok"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Domain'
        404:
          description: "Not Found"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
        400:
          description: "Bad Request"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'
  /custom-domains:
    get:
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Filtro para el dominio a buscar
          required: false
      operationId: "api.domains.obtener_todos"
      tags:
        - "Domains"
      summary: "Obtiene todos los custom domains del listado de domains"
      description: "Obtiene todos los custom domains"
      responses:
        200:
          description: "Ok"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Domain'
    post:
      operationId: "api.domains.crear"
      tags:
        - "Domains"
      summary: "Alta de un dominio custom"
      description: "Crea un nuevo dominio y lo agrega a la lista de custom domains"
      requestBody:
        required: true
        x-body-name: 'domain'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputDomain'
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        400:
          description: "Bad Request"
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Error'

# Declaración de tipos de datos
components:
  schemas:
    Domain:
      type: "object"
      properties:
        domain:
          type: "string"
        ip:
          type: "string"
        custom:
          type: "boolean"
    InputDomain:
      type: "object"
      properties:
        domain:
          type: "string"
        ip:
          type: "string"
    OnlyDomain:
      type: "object"
      properties:
        domain:
          type: "string"
    OnlyIp:
      type: "object"
      properties:
        ip:
          type: "string"
    Error:
      type: "object"
      properties:
        error:
          type: "string"